doc(derive): Update for value_parser

+++ examples/tutorial_derive/04_02_parse.rs
@@ -4,8 +4,8 @@ use clap::Parser;
 #[clap(author, version, about, long_about = None)]
 struct Cli {
     /// Network port to use
-    #[clap(parse(try_from_str))]
-    port: usize,
+    #[clap(value_parser = clap::value_parser!(u16).range(1..))]
+    port: u16,
 }
 
 fn main() {
